+++ packages/expo-video/ios/VideoModule.swift
@@ -320,6 +320,20 @@ public final class VideoModule: Module {
         player.ref.preventsDisplaySleepDuringVideoPlayback = keepScreenOnWhilePlaying
       }
 
+      Property("seekTolerance") { player -> SeekTolerance in
+        return player.seeker.seekTolerance
+      }
+      .set{ player, seekTolerance in
+        player.seeker.seekTolerance = seekTolerance
+      }
+
+      Property("scrubbingModeOptions") { player -> ScrubbingModeOptions in
+        return player.seeker.scrubbingModeOptions
+      }
+      .set {player, options in
+        player.seeker.scrubbingModeOptions = options
+      }
+
       Function("play") { player in
         player.ref.play()
       }
@@ -341,11 +355,11 @@ public final class VideoModule: Module {
       Function("seekBy") { (player, seconds: Double) in
         let newTime = player.ref.currentTime() + CMTime(seconds: seconds, preferredTimescale: .max)
 
-        player.ref.seek(to: newTime)
+        player.seeker.seek(to: newTime)
       }
 
       Function("replay") { player in
-        player.ref.seek(to: CMTime.zero)
+        player.seeker.seek(to: CMTime.zero)
         player.ref.play()
       }